tiff: fix compile warning

+++ backend/tiff/tiff2ps.c
@@ -1127,8 +1127,8 @@ PS_Lvl2page(TIFF2PSContext* ctx, TIFF* tif, uint32 w, uint32 h)
 	}
 	buf_data = (unsigned char *)_TIFFmalloc(chunk_size);
 	if (!buf_data) {
-		TIFFError(ctx->filename, "Can't alloc %i bytes for %s.",
-			chunk_size, tiled_image ? "tiles" : "strips");
+		TIFFError(ctx->filename, "Can't alloc %u bytes for %s.",
+			(uint) chunk_size, tiled_image ? "tiles" : "strips");
 		return(FALSE);
 	}
